+++ tests/Unit/ConfigTest.php
@@ -13,7 +13,7 @@ class ConfigTest extends LaravelTableTestCase
         // laravel-table
         $this->assertTrue(array_key_exists('classes', config('laravel-table')));
         $this->assertTrue(array_key_exists('icon', config('laravel-table')));
-        $this->assertTrue(array_key_exists('rows', config('laravel-table')));
+        $this->assertTrue(array_key_exists('value', config('laravel-table')));
         $this->assertTrue(array_key_exists('template', config('laravel-table')));
         // laravel-table.classes
         $this->assertTrue(array_key_exists('container', config('laravel-table.classes')));
@@ -22,6 +22,7 @@ class ConfigTest extends LaravelTableTestCase
         $this->assertTrue(array_key_exists('th', config('laravel-table.classes')));
         $this->assertTrue(array_key_exists('td', config('laravel-table.classes')));
         $this->assertTrue(array_key_exists('results', config('laravel-table.classes')));
+        $this->assertTrue(array_key_exists('disabled', config('laravel-table.classes')));
         // laravel-table.icon
         $this->assertTrue(array_key_exists('rowsNumber', config('laravel-table.icon')));
         $this->assertTrue(array_key_exists('sort', config('laravel-table.icon')));
@@ -33,14 +34,9 @@ class ConfigTest extends LaravelTableTestCase
         $this->assertTrue(array_key_exists('create', config('laravel-table.icon')));
         $this->assertTrue(array_key_exists('edit', config('laravel-table.icon')));
         $this->assertTrue(array_key_exists('destroy', config('laravel-table.icon')));
-        // laravel-table.rows
-        $this->assertTrue(array_key_exists('number', config('laravel-table.rows')));
-        $this->assertTrue(array_key_exists('disabled', config('laravel-table.rows')));
-        // laravel-table.rows.number
-        $this->assertTrue(array_key_exists('default', config('laravel-table.rows.number')));
-        $this->assertTrue(array_key_exists('selection', config('laravel-table.rows.number')));
-        // laravel-table.rows.classes
-        $this->assertTrue(array_key_exists('classes', config('laravel-table.rows.disabled')));
+        // laravel-table.value
+        $this->assertTrue(array_key_exists('rowsNumber', config('laravel-table.value')));
+        $this->assertTrue(array_key_exists('rowsNumberSelectionActivation', config('laravel-table.value')));
         // laravel-table.template
         $this->assertTrue(array_key_exists('table', config('laravel-table.template')));
         $this->assertTrue(array_key_exists('thead', config('laravel-table.template')));
@@ -50,7 +46,7 @@ class ConfigTest extends LaravelTableTestCase
 
     public function testCustomDefaultValueRowsNumber()
     {
-        config()->set('laravel-table.rows.number.default', 9999);
+        config()->set('laravel-table.value.rowsNumber', 9999);
         $this->createMultipleUsers(3);
         $this->routes(['users'], ['index', 'create', 'edit', 'destroy']);
         $table = (new Table)->model(User::class)->routes([