+++ views/bootstrap/tfoot.blade.php
@@ -2,10 +2,10 @@
     <tr {{ classTag($table->trClasses) }}>
         <td {{ classTag('bg-light', $table->tdClasses) }}
             {{ htmlAttributes($table->columnsCount() > 1 ? ['colspan' => $table->columnsCount()] : null) }}>
-            <div class="d-flex justify-content-between flex-wrap py-3">
+            <div class="d-flex justify-content-between flex-wrap py-2">
                 {{-- create button --}}
                 @if($table->isRouteDefined('create'))
-                    <div class="d-flex px-4 py-2 creation-container">
+                    <div class="d-flex align-items-center px-3 py-1 creation-container">
                         <a href="{{ $table->route('create') }}"
                            class="btn btn-success"
                            title="{{ __('laravel-table::laravel-table.create') }}">
@@ -15,11 +15,11 @@
                     </div>
                 @endif
                 {{-- navigation --}}
-                <div class="d-flex align-items-center px-4 py-2 navigation-container">
+                <div class="d-flex align-items-center px-3 py-1 navigation-container">
                     <div>{!! $table->navigationStatus() !!}</div>
                 </div>
                 {{-- pagination --}}
-                <div class="d-flex px-4 py-2 pagination-container">
+                <div class="d-flex align-items-center mb-n3 px-3 py-1 pagination-container">
                     {!! $table->list->links() !!}
                 </div>
             </div>